[ADD] Doctor test

+++ app/web_server/server/pharmaxcess_server/src/main/java/com/pharmaxcess_server/pharmaxcess_server/repository/DoctorRepository.java
@@ -26,7 +26,7 @@ public interface DoctorRepository extends JpaRepository<Doctor, Long> {
      * @throws IllegalArgumentException if id is {@code null}
      */
     @Query("SELECT o FROM Doctor o WHERE o.id = :id")
-    Optional<Doctor> findDoctorById(@Param("id") Long id);
+    Optional<Doctor> findDoctorById(@Param("id") Integer id);
 
     /**
      * Retrieves all doctors.
